+++ lib_board_support/api/boards/xk_voice_l71/board.h
@@ -1,12 +1,10 @@
-// Copyright 2024 XMOS LIMITED.
+// Copyright 2024-2025 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #ifndef __XK_VOICE_L71_BOARD_H__
 #define __XK_VOICE_L71_BOARD_H__
 
 #include <xccompat.h>
-#include "i2c.h"
-
 
 
 /** 
@@ -14,28 +12,34 @@
  * or an adjustable or fixed clock using the on-chip application PLL.
  */
 typedef enum {
-    CLK_FIXED,      /** Generate fixed MCLK from XCORE using APP_PLL */ 
-    CLK_EXTERNAL    /** Expect an externally provided MCLK  */
+    /** Generate fixed MCLK from XCORE using APP_PLL */ 
+    CLK_FIXED,
+    /** Expect an externally provided MCLK. Note you will need to set ENABLE_MCLK in xk_voice_l71_rpi_enable_t during xk_voice_l71_AudioHwInit() */
+    CLK_EXTERNAL
 } xk_voice_l71_mclk_modes_t;
 
 /** 
  * @brief Which DAC I2S data pin to use.
  */
 typedef enum {
-    DAC_DIN_PRI,   /** Use default DAC data pin */ 
-    DAC_DIN_SEC    /** Use secondary DAC data pin */
+    /** Use default DAC data pin (primary) */ 
+    DAC_DIN_PRI,
+    /** Use secondary DAC data pin (secondary)*/
+    DAC_DIN_SEC
 } xk_voice_l71_dac_pin_t;
 
 /** 
  * @brief Which of the I2C expander enable lines to set. These can be
  * ORed together. Setting these will connect the xcore chip to the 
- * raspberry PI connector
+ * raspberry PI expansion connector.
  */
 typedef enum {
-    NO_OE_PINS = 0x00,    /** All OEs disables */
-    MCLK_OE_PIN = 0x1,    /** Enable MCLK to/from the Pi connector */ 
-    SPI_OE_PIN = 0x2,     /** Enable SPI to/from the Pi connector */
-    I2S_OE_PIN = 0x4      /** Enable SPI to/from the Pi connector */
+    /** All OEs disabled */
+    ENABLE_NO_PINS  = 0x0,
+    ENABLE_MCLK     = 0x1,
+    ENABLE_SPI      = 0x2,
+    ENABLE_I2S      = 0x4,
+    ENABLE_INT      = 0x8
 } xk_voice_l71_rpi_enable_t;
 
 /**
@@ -45,9 +49,11 @@ typedef enum {
 typedef struct {
     /** xk_voice_l71_config_t::clk_mode See xk_voice_l71_mclk_modes_t for available clock mode options. */
     xk_voice_l71_mclk_modes_t clk_mode;
-    xk_voice_l71_rpi_enable_t oe_enable;
+    /** xk_voice_l71_config_t::oe_enables See xk_voice_l71_rpi_enable_t for available clock mode options. */
+    xk_voice_l71_rpi_enable_t oe_enables;
+    /** xk_voice_l71_config_t::dac_pin See xk_voice_l71_dac_pin_t for available clock mode options. */
     xk_voice_l71_dac_pin_t dac_pin;
-    char dac_is_clock_master;
+    /** The initial MCLK frequency in Hz to output before xk_voice_l71_AudioHwConfig() is called */
     unsigned default_mclk;
 } xk_voice_l71_config_t;
 
@@ -59,26 +65,16 @@ typedef struct {
  * @{
  */
 
-/** Command enumeration for channel based commands to I2C master server on other tile.
- */
-typedef enum
-{
-    AUDIOHW_CMD_REGWR,
-    AUDIOHW_CMD_REGRD,
-    AUDIOHW_CMD_EXIT
-} xk_voice_l71_audioHwCmd_t;
-
-
-/** Starts an I2C master server task. Must be started *before* the tile[1] xk_voice_l71_AudioHwInit calls. 
+/** Starts an I2C master server task. Must be started on tile[0] *before* the tile[1] xk_voice_l71_AudioHwInit calls. 
  * In the background this also starts a combinable channel to interface translation task
  * so the API may be used over a channel end however it still only occupies one thread.
- * May be exited after config by sending AUDIOHW_CMD_EXIT if dynamic configuration is not required.
+ * May be exited after config by calling xk_voice_l71_AudioHwRemoteKill() if dynamic configuration is not required.
  *
  *  \param   c    Server side of channel connecting I2C master server and HW config functions.
  */
 void xk_voice_l71_AudioHwRemote(chanend c);
 
-/** Initialises the client side channel for remote communications with I2C. Must be called on tile[1] *before* xk_voice_l71_AudioHwInit(). 
+/** Initialises the client side global channel end for remote communications with I2C. Must be called on tile[1] *before* xk_voice_l71_AudioHwInit(). 
  *
  *  \param   c    Client side of channel connecting I2C master server and HW config functions.
  */
@@ -100,6 +96,13 @@ void xk_voice_l71_AudioHwConfig(const REFERENCE_PARAM(xk_voice_l71_config_t, con
                                 unsigned sample_rate,
                                 unsigned mClk);
 
+/** Kills the remote I2C task. No further DAC or board config will be possible unless xk_voice_l71_AudioHwRemote() is restarted.
+ *
+ *  \param   c    Server side of channel connecting I2C master server and HW config functions.
+ */
+void xk_voice_l71_AudioHwRemoteKill(void);
+
 /**@}*/ // END: addtogroup xk_voice_l71
+
 
 #endif // __XK_VOICE_L71_BOARD_H__